WTS : added functions for vertical zooming + fixed bugs
git-svn-id: https://svn.opendtect.org/od/trunk@16373 a985855d-d384-4b45-ae33-31a410998c3c

+++ include/WellAttrib/welltieextractdata.h
@@ -32,7 +32,7 @@ public:
 			~WellTieExtractTrack() {};
 
 
-    StepInterval<float> timeintv_;
+    StepInterval<double> timeintv_;
 
     int                 nextStep();
     od_int64            totalNr() const		{ return timeintv_.nrSteps(); }
@@ -57,7 +57,7 @@ public:
 			~WellTieResampleLog() {};
 
 
-    StepInterval<float> timeintv_;
+    StepInterval<double> timeintv_;
 
     int                 nextStep();
     int           	colnr_;
